Organização estrutura do projeto

+++ scripts/meu-script.sh
@@ -1,45 +1,38 @@
+
 #!/bin/bash
 
 echo "Organizando projeto!"
 
-mkdir -p ../src ../tb ../include ../docs ../scripts
+mkdir -p ../src ../tb ../include ../docs
 
 # include (.vh)
-for file in ../*.vh; do
+for file in *.vh; do
     if [ -f "$file" ]; then
-        echo "Movendo $file para include/"
+        echo "Movendo $file para ../include/"
         mv -n "$file" ../include/
     fi
 done
 
-# scripts (.tcl, .do, .sh)
-for file in ../*.tcl ../*.do ../*.sh; do
-    if [ -f "$file" ]; then
-        echo "Movendo $file para scripts/"
-        mv -n "$file" ../scripts/
-    fi
-done
-
 # docs (.md, .txt)
-for file in ../*.md ../*.txt; do
+for file in *.md *.txt; do
     if [ -f "$file" ]; then
-        echo "Movendo $file para docs/"
+        echo "Movendo $file para ../docs/"
         mv -n "$file" ../docs/
     fi
 done
 
-# Testbench (.v com _tb ou test)
-for file in ../*.v; do
-    if [[ "$file" == *_tb.v || "$file" == *test* || "$file" == *-tb.v ]]; then
-        echo "Movendo $file para tb/"
+# Testbench
+for file in *.v; do
+    if [ -f "$file" ] && [[ "$file" == *_tb.v || "$file" == *test* || "$file" == *-tb.v ]]; then
+        echo "Movendo $file para ../tb/"
         mv -n "$file" ../tb/
     fi
 done
 
-# Arquivos restantes (.v → src)
-for file in ../*.v; do
+# resto dos .v
+for file in *.v; do
     if [ -f "$file" ]; then
-        echo "Movendo $file para src/"
+        echo "Movendo $file para ../src/"
         mv -n "$file" ../src/
     fi
 done
